+++ tests/test_builder.py
@@ -2,7 +2,7 @@
 import pytest
 
 # Local imports
-from uplink import builder, converter, client, exceptions, utils
+from uplink import builder, converter, hooks, exceptions, utils
 
 
 @pytest.fixture
@@ -15,35 +15,52 @@ def fake_service_cls(request_definition_builder, request_definition):
 
 @pytest.fixture
 def uplink_builder():
-    return builder.Builder(None)
+    return builder.Builder()
 
 
 class TestResponseConverter(object):
-    def test_handle_response(self, mocker, http_client_mock):
+    def test_handle_response(self, mocker, transaction_hook_mock):
         converter_ = mocker.Mock()
         input_value = "converted"
         converter_.return_value = input_value
-        rc = builder.ResponseConverter(http_client_mock, converter_)
+        rc = builder.ResponseConverter(transaction_hook_mock, converter_)
         assert rc.handle_response(None) is input_value
 
 
+class TestRequestHandler(object):
+
+    def test_fulfill(self, mocker, request_mock):
+        hook = mocker.Mock(spec=hooks.BaseTransactionHook)
+        request_mock.send.return_value = 1
+
+        request_handler = builder.RequestHandler(hook, 1, 2, 3)
+        value = request_handler.fulfill(request_mock)
+
+        hook.audit_request(1, 2, 3)
+        request_mock.add_callback.assert_called_with(hook.handle_response)
+        assert value == 1
+
+
 class TestRequestPreparer(object):
+
     def test_prepare_request(
             self,
-            http_client_mock,
             request_definition,
-            uplink_builder
+            uplink_builder,
+            transaction_hook_mock
     ):
         request = utils.Request("METHOD", "/example/path", {}, None)
-        uplink_builder.client = http_client_mock
+        uplink_builder.hook = transaction_hook_mock
         uplink_builder.base_url = "https://example.com"
         request_preparer = builder.RequestPreparer(
             uplink_builder, request_definition
         )
-        return_value = request_preparer.prepare_request(request)
-        assert return_value[0] == "METHOD"
-        assert return_value[1] == "https://example.com/example/path"
-        assert return_value[2] == {}
+        request_preparer.prepare_request(request)
+        transaction_hook_mock.audit_request.assert_called_with(
+            "METHOD",
+            "https://example.com/example/path",
+            {}
+        )
 
 
 class TestCallFactory(object):
@@ -57,23 +74,11 @@ class TestCallFactory(object):
             instance,
             request_preparer,
             request_definition)
-        assert isinstance(factory(*args, **kwargs), builder.Call)
+        assert factory(*args, **kwargs) is request_preparer.prepare_request.return_value
         request_definition.define_request.assert_called_with(
             request_builder, (instance,) + args, kwargs
         )
         assert request_builder.build.called
-
-
-class TestCall(object):
-    def test_return_type(self):
-        call = builder.Call(None, str)
-        assert call.return_type is str
-
-    def test_execute(self, mocker):
-        prepared_request = mocker.Mock(spec=client.PreparedRequest)
-        call = builder.Call(prepared_request, None)
-        call.execute()
-        prepared_request.send.assert_called_with()
 
 
 class TestBuilder(object):
@@ -83,8 +88,9 @@ class TestBuilder(object):
             converter.StandardConverterFactory
         )
 
-    def test_client_getter(self, uplink_builder):
-        assert isinstance(uplink_builder.client, client.BaseHttpClient)
+    def test_client_getter(self, uplink_builder, http_client_mock):
+        uplink_builder.client = http_client_mock
+        assert uplink_builder.client is http_client_mock
 
     def test_client_setter(self, uplink_builder, http_client_mock):
         uplink_builder.client = http_client_mock
@@ -101,27 +107,26 @@ class TestBuilder(object):
         factory = list(uplink_builder.converter_factories)[0]
         assert factory == converter_factory_mock
 
-    def test_build(self, fake_service_cls):
-        service = builder.Builder(fake_service_cls).build()
-        assert isinstance(service, fake_service_cls)
-
     def test_build_failure(self, fake_service_cls):
         exception = exceptions.InvalidRequestDefinition()
         fake_service_cls.builder.build.side_effect = exception
-        uplink = builder.Builder(fake_service_cls)
+        fake_service_cls.builder.__name__ = "builder"
+        uplink = builder.Builder()
         with pytest.raises(exceptions.UplinkBuilderError):
-            uplink.build()
+            uplink.build(fake_service_cls(), fake_service_cls.builder)
 
 
-def test_build(mocker, http_client_mock):
+def test_build(mocker, http_client_mock, fake_service_cls):
+    # Monkey-patch the Builder class.
     builder_cls_mock = mocker.Mock()
     builder_mock = mocker.Mock(spec=builder.Builder)
     builder_cls_mock.return_value = builder_mock
     mocker.patch.object(builder, "Builder", builder_cls_mock)
+
     builder.build(
         fake_service_cls,
         base_url="example.com",
-        http_client=http_client_mock
+        client=http_client_mock
     )
     assert builder_mock.base_url == "example.com"
     builder_mock.add_converter_factory.assert_called_with()