+++ tests/test_integration.py
@@ -14,24 +14,24 @@ def _get_url(url):
 
 
 @uplink.headers({"Accept": "application/vnd.github.v3.full+json"})
-class GitHubService(object):
+class GitHubService(uplink.Consumer):
 
     @uplink.get("/users/{user}/repos")
     def list_repos(self, user): pass
 
 
 @pytest.fixture
-def builder(http_client_mock):
-    builder_ = uplink.Builder(GitHubService)
-    builder_.base_url = BASE_URL
-    builder_.client = http_client_mock
-    return builder_
+def github_service_and_client(transaction_hook_mock):
+    return (
+        GitHubService(base_url=BASE_URL, hook=transaction_hook_mock),
+        transaction_hook_mock
+    )
 
 
-def test_list_repo(builder):
-    service = builder.build()
-    service.list_repos("prkumar").execute()
-    builder.client.audit_request.assert_called_with(
+def test_list_repo(github_service_and_client):
+    service, transaction_hook_mock = github_service_and_client
+    service.list_repos("prkumar")
+    transaction_hook_mock.audit_request.assert_called_with(
         "GET", _get_url("/users/prkumar/repos"), {
             "headers": {
                 "Accept": "application/vnd.github.v3.full+json"